ERR: Attempt to fix more Curl warnings

+++ Source/CTest/Curl/getdate.c
@@ -1796,7 +1796,7 @@ LookupWord (yylval, buff)
   /* Make it lowercase. */
   for (p = buff; *p; p++)
     if (ISUPPER ((unsigned int)(*p)))
-      *p = (char)tolower (*p);
+      *p = (char)tolower ((unsigned int)(*p));
 
   if (strcmp (buff, "am") == 0 || strcmp (buff, "a.m.") == 0)
     {
